feat: Implement expense management functionality with service, controller, and DTOs

+++ src/main/java/mr/limpios/smart_divide_backend/infraestructure/dto/AddExpenseDTO.java
@@ -1,13 +1,12 @@
 package mr.limpios.smart_divide_backend.infraestructure.dto;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 import mr.limpios.smart_divide_backend.domain.models.DivisionType;
 public record AddExpenseDTO(
         String type,
         String description,
-        BigDecimal amount,
+        double amount,
         String evidenUrl,
         DivisionType divisionType,
         List<AddExpenseDebtorsDTO> balances
